Add the update function behind the Edit / Submit button.

+++ app/assets/javascripts/components/_item_table.js.jsx
@@ -3,7 +3,7 @@ class ItemTable extends React.Component {
   render() {
     var items = this.props.items.map((item) => {
       return (
-          <Item item={item}
+          <Item item={item} key={item.id}
             handleDelete={this._handleDeleteClick.bind(this, item.id)}
             handleUpdate={this._onUpdate.bind(this)} />       
       )
@@ -25,11 +25,15 @@ class ItemTable extends React.Component {
 
   _handleDeleteClick(id) {
     console.log('delete item clicked');
-    this.props.handleDelete(id);
+    console.log(id);
+    if(confirm("Are you sure to delete this item?")) {
+      this.props.handleDelete(id);
+    }
   }
 
   _onUpdate(item) {
     console.log('update item in table');
+    console.log(item);
     this.props.onUpdate(item);
   } 
 }
